pyproject.toml
[project]
name = "my-family-bot"
version = "0.1.0"
description = "Family Tree AI Assistant — GEDCOM → Neo4j → Multi-Agent → Web + WhatsApp"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "neo4j>=5.24.0",
    "python-gedcom>=1.0.0",
    "twilio>=9.3.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
